+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [unreleased]
 ### Added
-- Optional auth protection on overview and report endpoint, when env variables `JWKS_URL` and `AUDIENCE` are specified
-- Tests for authorized endpoints
+- Optional Authorization protection on overview and report endpoints. When env variables `JWKS_URL` and `AUDIENCE` are specified, `access_token` will be collected from request cookies
+- Optional Authorization protection on coverage endpoints. When env variables `JWKS_URL` and `AUDIENCE` are specified, `access_token` will be collected from request headers {"Authorization": "Bearer <token>"}
+- Tests for endpoints protected by Authorization
 
 ## [3.5.1]
 ### Added